+++ ctermui/ctermui_pencil.h
@@ -9,13 +9,13 @@ typedef char*** screen_buffer;
 #define CTERMUI_HORIZONTAL_LINE     '-'
 #define CTERMUI_VERTICAL_LINE       '|'
 
-size_t ctermui_pencil_draw_char(screen_buffer b,
+int ctermui_pencil_draw_char(screen_buffer b,
                              size_t x,
                              size_t y,
                              char c,
                              size_t fg_color,
                              size_t bg_color);
-size_t ctermui_pencil_draw_line(screen_buffer b,
+int ctermui_pencil_draw_line(screen_buffer b,
                              size_t orientation,
                              size_t x,
                              size_t y,
@@ -23,26 +23,26 @@ size_t ctermui_pencil_draw_line(screen_buffer b,
                              char c,
                              size_t color,
                              size_t bg_color);
-size_t ctermui_pencil_draw_rect(screen_buffer b,
+int ctermui_pencil_draw_rect(screen_buffer b,
                              size_t x,
                              size_t y,
                              size_t width,
                              size_t height,
                              size_t color,
                              size_t bg_color);
-size_t ctermui_pencil_draw_text(screen_buffer b,
+int ctermui_pencil_draw_text(screen_buffer b,
                              size_t x,
                              size_t y,
                              char* text,
                              size_t color,
                              size_t bg_color);
-size_t ctermui_pencil_solid_background(screen_buffer b,
+int ctermui_pencil_solid_background(screen_buffer b,
                                     size_t x,
                                     size_t y,
                                     size_t width,
                                     size_t height,
                                     size_t color);
-size_t ctermui_pencil_bucket(screen_buffer b,
+int ctermui_pencil_bucket(screen_buffer b,
                           size_t x,
                           size_t y,
                           size_t width,
